% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxt5.R
\docType{package}
\name{Maxt5}
\alias{Maxt5}
\title{Find test statistic value from a given data and corresponding critical value}
\description{
This package finds test statistic value and
critical value of Max-T test for testing equality of means against
ordered alternatives in one-way ANOVA model with heteroscedasticity
when number of groups is 5.
}
\author{
Anjana Mondal \email{mondal.anjana25@gmail.com}
}
