% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxt5crit.R
\name{Maxt5crit}
\alias{Maxt5crit}
\title{Find test statistic value from a given data and corresponding critical value}
\usage{
Maxt5crit(g1, g2, g3, g4, g5, alpha)
}
\arguments{
\item{g1}{a real vector}

\item{g2}{a real vector}

\item{g3}{a real vector}

\item{g4}{a real vector}

\item{g5}{a real vector}

\item{alpha}{real number between 0 and 1 called significance level}
}
\value{
numeric vector
}
\description{
More detailed description
}
\examples{
g1=c(1,2,3,2,0,1,4,2,3)
g2=c(1,3,1,2,5,2,1,7,8,1)
g3=c(1,1,2,4,5,6,2,5,7,2)
g4=c(1,2,1,3,1,5,1,3,2,3,0,2)
g5=c(2,4,2,7,3,7,1,2,3,5,7)

Maxt5crit(g1,g2,g3,g4,g5,0.05)


}
